Pagamentos
Frond end e backend funcionando perfeitamente e aceitando os pagamentos. Falta criar a tabela de pagamentos

+++ server/src/main/java/com/ecommerce/entities/dto/pagamentoDTO.java
@@ -1,5 +1,7 @@
 package com.ecommerce.entities.dto;
 
-public record pagamentoDTO (String idPedido, double totalPedido, String userEmail, String userFullName, String cpf ){
+import java.math.BigDecimal;
+
+public record pagamentoDTO (String idPedido, BigDecimal totalPedido, String userEmail, String userFullName, String cpf ){
 
 }
